+++ kitchenpos-domain/order-table/src/main/java/kitchenpos/ordertable/application/OrderTableEventListener.java
@@ -10,8 +10,8 @@ import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableUngroupValidator;
 import kitchenpos.ordertable.repository.OrderTableRepository;
 import kitchenpos.tablegroup.domain.TableGroup;
-import kitchenpos.tablegroup.dto.TableGroupGroupEvent;
-import kitchenpos.tablegroup.dto.TableGroupUngroupEvent;
+import kitchenpos.tablegroup.event.TableGroupGroupEvent;
+import kitchenpos.tablegroup.event.TableGroupUngroupEvent;
 import kitchenpos.tablegroup.repository.TableGroupRepository;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
