+++ kitchenpos-domain/order-table/src/test/java/kitchenpos/ordertable/application/OrderTableEventListenerTest.java
@@ -16,8 +16,8 @@ import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableUngroupValidator;
 import kitchenpos.ordertable.repository.OrderTableRepository;
 import kitchenpos.tablegroup.domain.TableGroup;
-import kitchenpos.tablegroup.dto.TableGroupGroupEvent;
-import kitchenpos.tablegroup.dto.TableGroupUngroupEvent;
+import kitchenpos.tablegroup.event.TableGroupGroupEvent;
+import kitchenpos.tablegroup.event.TableGroupUngroupEvent;
 import kitchenpos.tablegroup.repository.TableGroupRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
