+++ kitchenpos-domain/table-group/src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import kitchenpos.common.exception.KitchenPosException;
 import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.tablegroup.dto.TableGroupCreateRequest;
-import kitchenpos.tablegroup.dto.TableGroupGroupEvent;
+import kitchenpos.tablegroup.event.TableGroupGroupEvent;
 import kitchenpos.tablegroup.dto.TableGroupResponse;
-import kitchenpos.tablegroup.dto.TableGroupUngroupEvent;
+import kitchenpos.tablegroup.event.TableGroupUngroupEvent;
 import kitchenpos.tablegroup.repository.TableGroupRepository;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
