+++ kitchenpos-domain/table-group/src/main/java/kitchenpos/tablegroup/event/TableGroupGroupEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.dto;
+package kitchenpos.tablegroup.event;
 
 import java.util.List;
