+++ kitchenpos-domain/table-group/src/main/java/kitchenpos/tablegroup/event/TableGroupUngroupEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.dto;
+package kitchenpos.tablegroup.event;
 
 public class TableGroupUngroupEvent {
